feat(comps): Add ColorBox
