--- App.js
@@ -1,61 +1,0 @@
-import React from 'react';
-import { View, Text, SafeAreaView, StyleSheet } from 'react-native';
-
-const App = () => {
-  return (
-    <SafeAreaView style={styles.safeArea}>
-      <Text style={styles.heading}>
-        Here are some boxes of different colours
-      </Text>
-      <View style={[styles.container, styles.cyan]}>
-        <Text style={styles.text}>Cyan: #2aa198</Text>
-      </View>
-      <View style={[styles.container, styles.blue]}>
-        <Text style={styles.text}>Blue: #268bd2</Text>
-      </View>
-      <View style={[styles.container, styles.magenta]}>
-        <Text style={styles.text}>Magenta: #d33682</Text>
-      </View>
-      <View style={[styles.container, styles.orange]}>
-        <Text style={styles.text}>Orange: #cb4b16</Text>
-      </View>
-    </SafeAreaView>
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    padding: 10,
-    borderRadius: 3,
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginBottom: 10,
-  },
-  heading: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    marginBottom: 10,
-  },
-  safeArea: {
-    paddingTop: 50,
-    paddingHorizontal: 10,
-  },
-  text: {
-    fontWeight: 'bold',
-    color: 'white',
-  },
-  cyan: {
-    backgroundColor: '#2aa198',
-  },
-  blue: {
-    backgroundColor: '#268bd2',
-  },
-  magenta: {
-    backgroundColor: '#d33682',
-  },
-  orange: {
-    backgroundColor: '#cb4b16',
-  },
-});
-
-export default App;
